Recognizer buttons
Since size and tone of color of music sheets in the world is varied, there are buttons for custom setting. With BRH, user can increase/decrease the  interval accepted as black color. With OS, the size of oval can be changed. With FN, user can increase/decrease the interval accepted as a note.