Upload image function
It would be weird to ask an user who doesn't know anything about this program to choose a file at the time the user opens this program. Therefore instead of showing filechooser, I let user to look what's in this program and be able to read manual. And then let the user upload a file. Also even after one musicsheet is scanned, user can upload another image through upload button.